requirements.txt
torch==2.2.1
torchvision==0.17.1
torchaudio==2.2.1
diffusers==0.29.0
transformers==4.38.1
accelerate==0.28.0
datasets==2.18.0
peft==0.11.1
opencv-contrib-python==4.9.0.80
Pillow==10.2.0
imageio==2.34.0
imageio-ffmpeg==0.4.9
numpy==1.26.4
scipy==1.12.0
pandas==2.2.1
tqdm==4.66.2
pyyaml==6.0.1
huggingface-hub==0.21.3
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
httpx==0.27.0
loguru==0.7.2
rich==13.7.1
wandb==0.16.4

